+++ OKX.Api/Grid/Responses/OkxGridContractOrderUpdate.cs
@@ -324,5 +324,5 @@ public record OkxGridContractOrderUpdate
     /// Push time of algo grid information
     /// </summary>
     [JsonIgnore]
-    public DateTime? PushTimes => PushTimestamp?.ConvertFromMilliseconds();
+    public DateTime? PushTime => PushTimestamp?.ConvertFromMilliseconds();
 }