+++ OKX.Api/Public/Responses/OkxPublicDiscountInfo.cs
@@ -12,10 +12,14 @@ public record OkxPublicDiscountInfo
     public string Currency { get; set; } = string.Empty;
 
     /// <summary>
-    /// Platform level collateralized borrow restriction
+    /// Platform level collateral restriction status
+    /// 0: The restriction is not enabled.
+    /// 1: The restriction is not enabled. But the crypto is close to the platform's collateral limit.
+    /// 2: The restriction is enabled.This crypto can't be used as margin for your new orders. This may result in failed orders. But it will still be included in the account's adjusted equity and doesn't impact margin ratio.
+    /// Refer to Introduction to the platform collateralized borrowing limit for more details.
     /// </summary>
-    [JsonProperty("collateralRestrict")]
-    public bool? IsCollateralRestricted { get; set; }
+    [JsonProperty("colRes")]
+    public OkxPublicCollateralRestrictionStatus CollateralRestrictionStatus { get; set; }
 
     /// <summary>
     /// Amount
@@ -23,12 +27,6 @@ public record OkxPublicDiscountInfo
     [JsonProperty("amt")]
     public decimal Amount { get; set; }
 
-    /// <summary>
-    /// Details
-    /// </summary>
-    [JsonProperty("discountInfo")]
-    public List<OkxPublicDiscountInfoData> Info { get; set; } = [];
-    
     /// <summary>
     /// Minimum discount rate when it exceeds the maximum amount of the last tier.
     /// </summary>
@@ -38,32 +36,8 @@ public record OkxPublicDiscountInfo
     /// <summary>
     /// Details
     /// </summary>
-    [JsonProperty("discountInfo")]
+    [JsonProperty("details")]
     public List<OkxPublicDiscountInfoDetails> Details { get; set; } = [];
-}
-
-/// <summary>
-/// OKX Public Discount Info Detail
-/// </summary>
-public record OkxPublicDiscountInfoData
-{
-    /// <summary>
-    /// Discount Rate
-    /// </summary>
-    [JsonProperty("discountRate")]
-    public decimal DiscountRate { get; set; }
-
-    /// <summary>
-    /// Maximum Amount
-    /// </summary>
-    [JsonProperty("maxAmt")]
-    public decimal MaximumAmount { get; set; }
-
-    /// <summary>
-    /// Minimum Amount
-    /// </summary>
-    [JsonProperty("minAmt")]
-    public decimal MinimumAmount { get; set; }
 }
 
 /// <summary>