+++ OKX.Api/Public/Responses/OkxPublicTradeAggregate.cs
@@ -10,4 +10,10 @@ public record OkxPublicTradeAggregate : OkxPublicTrade
     /// </summary>
     [JsonProperty("count")]
     public int Count { get; set; }
+
+    /// <summary>
+    /// Sequence ID
+    /// </summary>
+    [JsonProperty("seqId")]
+    public long SequenceId { get; set; }
 }